--- src/main/java/project/back/entitiy/CartItem.java
@@ -1,38 +1,0 @@
-package project.back.entitiy;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-
-@Entity
-@Table(name = "cart_item")
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class CartItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "cart_id", nullable = false, unique = true)
-    private Long itemId;
-
-    @Column(name = "quantity")
-    private Long quantity;
-
-
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="cart_id")
-    private Cart cart;
-
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="product_id")
-    private Product product;
-
-}
