+++ src/main/java/project/back/repository/memberrepository/MemberRepository.java
@@ -1,7 +1,7 @@
 package project.back.repository.memberrepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import project.back.entitiy.Member;
+import project.back.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member,Long>, MemberRepositoryOther {
 }
