+++ src/main/java/project/back/repository/memberrepository/MemberRepositoryOtherImpl.java
@@ -4,8 +4,7 @@ import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import project.back.dto.MemberDto;
-import project.back.entitiy.Member;
-import project.back.repository.memberrepository.MemberRepositoryOther;
+import project.back.entity.Member;
 
 import java.time.LocalDateTime;
 import java.util.List;
